fix: remove special _json_key handling
The approach of base64 encoding the json-like structure appears to work
fine. The shape of the moving parts (username, password) was largely reverse
engineered by:

1. Generating a service account with gcr access and saving to `key.json`.
2. `cat key.json | docker login -u _json_key --password-stdin https://gcr.io`
3. `cat `~/.docker/config.json` should now contain somethin like:```
    "gcr.io" : {
      "auth" : "..."
      }

The password should typically itself be a json-encoded service account
i.e. the contents of `key.json` above. We've replaced the parent commits
sprintf json-encoding by using the upstream types.

Note: we're panicking on the error here vs. propagating it. This is
partially laziness, the perceived unlikihood, and finally the desire to
keep this change set as straightforward as possible.
```